add database sinhronization in duration choose

+++ app/src/main/java/com/example/movieselector/mainmenu/models/Duration.kt
@@ -1,9 +1,9 @@
 package com.example.movieselector.mainmenu.models
 
 class Duration(
-    val hour: Short = 0,
-    val minute: Short = 0,
-    val error: Short = 0,
+    val hour: Int = 0,
+    val minute: Int = 0,
+    val error: Int = 0,
 ) {
     override fun toString(): String {
         return "${hour}h, ${minute}m, ${error}error"
